Removed old scheduler.py
